# back/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Inventory",
        "description": "Documentação da API Inventory gerada automaticamente via Swagger.",
        "contact": {
            "email": "victor.hugo@bizsys.com.br"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/asset-types": {
            "get": {
                "tags": [
                    "Tipos de Ativo"
                ],
                "summary": "Lista tipos de ativos",
                "description": "Retorna uma lista paginada de tipos de ativos.",
                "operationId": "b0ef8566e9fabc14db41157d6778fc26",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista paginada de tipos de ativos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Tipos de Ativo"
                ],
                "summary": "Cria um novo tipo de ativo",
                "description": "Cria um novo tipo de ativo.",
                "operationId": "9d5bc403bcbaaaa2a967e2792bdf61f1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Notebook"
                                    },
                                    "observation": {
                                        "type": "string",
                                        "example": "Equipamento de TI"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tipo de ativo criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação"
                    },
                    "500": {
                        "description": "Ocorreu um erro inesperado ao processar sua solicitação. Tente novamente mais tarde."
                    }
                }
            }
        },
        "/asset-types/{id_asset_type}": {
            "get": {
                "tags": [
                    "Tipos de Ativo"
                ],
                "summary": "Exibe um tipo de ativo",
                "description": "Retorna os dados de um tipo de ativo pelo ID.",
                "operationId": "0502a3dffcab908b01a6504b9bf5559d",
                "parameters": [
                    {
                        "name": "id_asset_type",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tipo de ativo encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tipo de ativo não encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Tipos de Ativo"
                ],
                "summary": "Atualiza um tipo de ativo",
                "description": "Atualiza os dados de um tipo de ativo.",
                "operationId": "7f2bb0eb9f43b06a0ac5d9c8e341f975",
                "parameters": [
                    {
                        "name": "id_asset_type",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Notebook"
                                    },
                                    "observation": {
                                        "type": "string",
                                        "example": "Equipamento de TI"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tipo de ativo atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação"
                    },
                    "404": {
                        "description": "Tipo de ativo não encontrado"
                    },
                    "500": {
                        "description": "Ocorreu um erro inesperado ao processar sua solicitação. Tente novamente mais tarde."
                    }
                }
            },
            "delete": {
                "tags": [
                    "Tipos de Ativo"
                ],
                "summary": "Remove um tipo de ativo",
                "description": "Remove (soft delete) um tipo de ativo.",
                "operationId": "9f40cfa912bfa7ce97494b9fd41edad5",
                "parameters": [
                    {
                        "name": "id_asset_type",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tipo de ativo removido com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tipo de ativo não encontrado"
                    },
                    "500": {
                        "description": "Ocorreu um erro inesperado ao processar sua solicitação. Tente novamente mais tarde."
                    }
                }
            }
        },
...
    },
...
}